Added clang-tidy; made code clang-tidy conformant.

+++ src/factor/factor.cc
@@ -2,14 +2,14 @@
 
 namespace factor {
 
-std::vector<long> factor(long x) {
+std::vector<int64_t> factor(int64_t x) {
   if (x == 1) {
     return {1};
   }
 
-  std::vector<long> factors;
+  std::vector<int64_t> factors;
   while (x != 1) {
-    for (int i = 2; i <= x; ++i) {
+    for (int64_t i = 2; i <= x; ++i) {
       if (x % i == 0) {
         x /= i;
         factors.push_back(i);
